requirements.txt
fastapi
uvicorn
PyJWT
requests
pydantic
cryptography
